Update README.md
An online shopping app project is a software application designed to provide a platform for users to browse, purchase, and manage products or services from various sellers. This type of project encompasses several key components and functionalities, all of which aim to create a seamless, efficient, and user-friendly shopping experience.